Refactor of the questionnaires/rest_framework/views/serializers.py
