+++ api/app/signals/apps/questionnaires/rest_framework/serializers/public/__init__.py
@@ -1,0 +1,2 @@
+# SPDX-License-Identifier: MPL-2.0
+# Copyright (C) 2021 Gemeente Amsterdam
