+++ api/app/signals/apps/questionnaires/rest_framework/serializers/public/answers.py
@@ -1,0 +1,59 @@
+# SPDX-License-Identifier: MPL-2.0
+# Copyright (C) 2021 Gemeente Amsterdam
+from datapunt_api.rest import DisplayField, HALSerializer
+from rest_framework.exceptions import ValidationError
+
+from signals.apps.questionnaires.models import Answer, Questionnaire, Session
+from signals.apps.questionnaires.rest_framework.fields import (
+    EmptyHyperlinkedIdentityField,
+    UUIDRelatedField
+)
+from signals.apps.questionnaires.services.utils import get_session_service
+
+
+class PublicAnswerSerializer(HALSerializer):
+    serializer_url_field = EmptyHyperlinkedIdentityField
+
+    _display = DisplayField()
+
+    session = UUIDRelatedField(uuid_field='uuid', queryset=Session.objects.retrieve_valid_sessions(), required=False)
+    questionnaire = UUIDRelatedField(uuid_field='uuid', queryset=Questionnaire.objects.active(), required=False)
+
+    class Meta:
+        model = Answer
+        fields = (
+            '_links',
+            '_display',
+            'payload',
+            'session',
+            'questionnaire',
+            'created_at',
+        )
+        read_only_fields = (
+            'created_at',
+        )
+
+    def validate(self, attrs):
+        attrs = super(PublicAnswerSerializer, self).validate(attrs=attrs)
+
+        if 'session' in attrs and 'questionnaire' in attrs:
+            raise ValidationError('session and questionnaire cannot be used both!')
+        elif 'session' not in attrs and 'questionnaire' not in attrs:
+            raise ValidationError('Either the session or questionnaire is mandatory!')
+
+        return attrs
+
+    def create(self, validated_data):
+        question = self.context['question']
+        payload = validated_data.pop('payload')
+
+        if 'session' in validated_data:
+            session = validated_data.pop('session')
+            session_service = get_session_service(session)
+        else:
+            questionnaire = validated_data.pop('questionnaire')
+            session = Session.objects.create(questionnaire=questionnaire)
+            session_service = get_session_service(session)
+
+        session_service.refresh_from_db()
+        return session_service.create_answer(payload, question)
